+++ sdk/identity/azure_identity/src/credentials/imds_managed_identity_credentials.rs
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-use crate::{credentials::cache::TokenCache, TokenCredentialOptions};
+use crate::{credentials::cache::TokenCache, env::Env, TokenCredentialOptions, UserAssignedId};
 use azure_core::{
     credentials::{AccessToken, Secret, TokenCredential},
     error::{http_response_from_body, Error, ErrorKind},
@@ -31,6 +31,16 @@ pub enum ImdsId {
     MsiResId(String),
 }
 
+impl From<UserAssignedId> for ImdsId {
+    fn from(user_assigned_id: UserAssignedId) -> Self {
+        match user_assigned_id {
+            UserAssignedId::ClientId(client_id) => ImdsId::ClientId(client_id),
+            UserAssignedId::ObjectId(object_id) => ImdsId::ObjectId(object_id),
+            UserAssignedId::ResourceId(resource_id) => ImdsId::MsiResId(resource_id),
+        }
+    }
+}
+
 /// Attempts authentication using a managed identity that has been assigned to the deployment environment.
 ///
 /// This authentication type works in Azure VMs, App Service and Azure Functions applications, as well as the Azure Cloud Shell
@@ -45,6 +55,7 @@ pub(crate) struct ImdsManagedIdentityCredential {
     secret_env: String,
     id: ImdsId,
     cache: TokenCache,
+    env: Env,
 }
 
 impl ImdsManagedIdentityCredential {
@@ -65,6 +76,7 @@ impl ImdsManagedIdentityCredential {
             secret_env: secret_env.to_owned(),
             id,
             cache: TokenCache::new(),
+            env: options.env().clone(),
         }
     }
 
@@ -90,7 +102,7 @@ impl ImdsManagedIdentityCredential {
 
         req.insert_header("metadata", "true");
 
-        let msi_secret = std::env::var(&self.secret_env);
+        let msi_secret = self.env.var(&self.secret_env);
         if let Ok(val) = msi_secret {
             req.insert_header(self.secret_header.clone(), val);
         };